Add warning to missing twitter resolver back
