+++ internal/resolvers/twitter/resolver.go
@@ -2,66 +2,15 @@ package twitter
 
 import (
 	"context"
-	"errors"
 	"net/http"
 	"net/url"
-	"regexp"
 	"strings"
-	"text/template"
 	"time"
 
 	"github.com/Chatterino/api/pkg/cache"
 	"github.com/Chatterino/api/pkg/config"
 	"github.com/Chatterino/api/pkg/resolver"
 	"github.com/Chatterino/api/pkg/utils"
-)
-
-const (
-	tweetTooltip = `<div style="text-align: left;">
-<b>{{.Name}} (@{{.Username}})</b>
-<span style="white-space: pre-wrap; word-wrap: break-word;">
-{{.Text}}
-</span>
-<span style="color: #808892;">{{.Likes}} likes&nbsp;•&nbsp;{{.Retweets}} retweets&nbsp;•&nbsp;{{.Timestamp}}</span>
-</div>
-`
-
-	twitterUserTooltip = `<div style="text-align: left;">
-<b>{{.Name}} (@{{.Username}})</b>
-<span style="white-space: pre-wrap; word-wrap: break-word;">
-{{.Description}}
-</span>
-<span style="color: #808892;">{{.Followers}} followers</span>
-</div>
-`
-)
-
-var (
-	tweetRegexp       = regexp.MustCompile(`(?i)\/.*\/status(?:es)?\/([^\/\?]+)`)
-	twitterUserRegexp = regexp.MustCompile(`(?i)twitter\.com\/([^\/\?\s]+)(\/?$|(\?).*)`)
-
-	/* These routes refer to non-user pages. If the capture group in twitterUserRegexp
-	   matches any of these names, we must not resolve it as a Twitter user link.
-
-	   The pages are listed alphabetically. They were sourced by simply looking around the
-	   Twitter web page. AFAIK, there is no resource describing these "special" routes.
-	*/
-	nonUserPages = utils.SetFromSlice([]interface{}{
-		"compose",
-		"explore",
-		"home",
-		"logout",
-		"messages",
-		"notifications",
-		"search",
-		"settings",
-		"tos",
-		"privacy",
-	})
-
-	tweetTooltipTemplate = template.Must(template.New("tweetTooltip").Parse(tweetTooltip))
-
-	twitterUserTooltipTemplate = template.Must(template.New("twitterUserTooltip").Parse(twitterUserTooltip))
 )
 
 type TwitterResolver struct {
@@ -121,11 +70,7 @@ func (r *TwitterResolver) Run(ctx context.Context, url *url.URL, req *http.Reque
 	return resolver.NoLinkInfoFound, nil
 }
 
-func NewTwitterResolver(ctx context.Context, cfg config.APIConfig) (*TwitterResolver, error) {
-	if cfg.TwitterBearerToken == "" {
-		return nil, errors.New("twitter-bearer-token is missing, won't do special responses for Twitter")
-	}
-
+func NewTwitterResolver(ctx context.Context, cfg config.APIConfig) *TwitterResolver {
 	tweetLoader := &TweetLoader{
 		bearerKey: cfg.TwitterBearerToken,
 	}
@@ -139,14 +84,5 @@ func NewTwitterResolver(ctx context.Context, cfg config.APIConfig) (*TwitterReso
 		userCache:  cache.NewPostgreSQLCache(ctx, cfg, "twitter:user", resolver.NewResponseMarshaller(userLoader), 24*time.Hour),
 	}
 
-	return r, nil
-}
-
-func Initialize(ctx context.Context, cfg config.APIConfig, resolvers *[]resolver.Resolver) {
-	resolver, err := NewTwitterResolver(ctx, cfg)
-	if err != nil {
-		return
-	}
-
-	*resolvers = append(*resolvers, resolver)
+	return r
 }
